+++ android/src/main/java/de/jonasbark/stripepayment/StripeDialog.kt
@@ -2,14 +2,14 @@ package de.jonasbark.stripepayment
 
 import android.os.Bundle
 import com.google.android.material.snackbar.Snackbar
-import androidx.fragment.app.DialogFragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.stripe.android.ApiResultCallback
+import com.stripe.android.PaymentConfiguration
 import com.stripe.android.SourceCallback
 import com.stripe.android.Stripe
-import com.stripe.android.model.Source
-import com.stripe.android.model.SourceParams
+import com.stripe.android.model.*
 import com.stripe.android.view.CardMultilineWidget
 import java.lang.Exception
 
@@ -28,8 +28,10 @@ class StripeDialog : androidx.fragment.app.DialogFragment() {
 
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         return inflater.inflate(R.layout.activity_stripe, container)
     }
 
@@ -40,7 +42,7 @@ class StripeDialog : androidx.fragment.app.DialogFragment() {
         val title = arguments?.getString("title", "Add Source")
         dialog.setTitle(title)
 
-        view?.findViewById<View>(R.id.buttonSave)?.setOnClickListener {
+        view.findViewById<View>(R.id.buttonSave)?.setOnClickListener {
             getToken()
         }
 
@@ -56,7 +58,8 @@ class StripeDialog : androidx.fragment.app.DialogFragment() {
     var tokenListener: ((String) -> (Unit))? = null
 
     private fun getToken() {
-        val mCardInputWidget = view?.findViewById<View>(R.id.card_input_widget) as CardMultilineWidget
+        val mCardInputWidget =
+            view?.findViewById<View>(R.id.card_input_widget) as CardMultilineWidget
 
         if (mCardInputWidget.validateAllFields()) {
 
@@ -66,32 +69,45 @@ class StripeDialog : androidx.fragment.app.DialogFragment() {
                 view?.findViewById<View>(R.id.buttonSave)?.visibility = View.GONE
 
                 val publishableKey = arguments?.getString("publishableKey", null) ?: ""
+                PaymentConfiguration.init(publishableKey)
 
-                val stripe = Stripe(activity!!, publishableKey)
-                stripe.createSource(SourceParams.createCardParams(card), object : SourceCallback {
-                    override fun onSuccess(source: Source) {
-                        view?.findViewById<View>(R.id.progress)?.visibility = View.GONE
-                        view?.findViewById<View>(R.id.buttonSave)?.visibility = View.GONE
+                val paymentMethodParamsCard = card.toPaymentMethodParamsCard()
+                val paymentMethodCreateParams = PaymentMethodCreateParams.create(
+                    paymentMethodParamsCard,
+                    PaymentMethod.BillingDetails.Builder().build()
+                )
 
-                        val sourceId = source.id ?: throw RuntimeException("Missing Source ID")
-                        tokenListener?.invoke(sourceId)
-                        dismiss()
-                    }
+                val stripe = Stripe(activity!!, PaymentConfiguration.getInstance().publishableKey)
 
-                    override fun onError(error: Exception) {
-                        view?.findViewById<View>(R.id.progress)?.visibility = View.GONE
-                        view?.findViewById<View>(R.id.buttonSave)?.visibility = View.VISIBLE
-                        view?.let {
-                            Snackbar.make(it, error!!.localizedMessage, Snackbar.LENGTH_LONG).show()
+                stripe.createPaymentMethod(
+                    paymentMethodCreateParams,
+                    object : ApiResultCallback<PaymentMethod> {
+                        override fun onSuccess(result: PaymentMethod) {
+                            view?.findViewById<View>(R.id.progress)?.visibility = View.GONE
+                            view?.findViewById<View>(R.id.buttonSave)?.visibility = View.GONE
+
+                            if (result.id != null) {
+                                tokenListener?.invoke(result.id!!)
+                                dismiss()
+                            }
                         }
-                    }
 
-                })
+                        override fun onError(error: Exception) {
+                            view?.findViewById<View>(R.id.progress)?.visibility = View.GONE
+                            view?.findViewById<View>(R.id.buttonSave)?.visibility = View.VISIBLE
+                            view?.let {
+                                Snackbar.make(it, error.localizedMessage, Snackbar.LENGTH_LONG)
+                                    .show()
+                            }
+                        }
+
+                    })
+
             }
-        }
-        else {
+        } else {
             view?.let {
-                Snackbar.make(it, "The card info you entered is not correct", Snackbar.LENGTH_LONG).show()
+                Snackbar.make(it, "The card info you entered is not correct", Snackbar.LENGTH_LONG)
+                    .show()
             }
         }